Compile 404.html and add Makefile
GitHub Pages requires /404.html to be a *raw* HTML file (i.e., not a
Jekyll-compiled one: https://help.github.com/articles/custom-404-pages).
As such, /_404.html is now the template for 404.html, while 404.html is
the compiled version (`make update-404` will do the build-and-copy,
additionally adding a build timestamp to the compiled file).